Handle connecting to Metamask and displaying register modal
This commit updates the logic of how the Card component handles the
various steps in connecting to a user's wallet and prompting them to
register.

When a user's currently selected Metamask account is not connected, we
display the "Connect wallet..." prompt. Upon clicking the button,
they'll be prompted to connect, and upon connecting, will be prompted
with the registration modal.

Previously, if the user clicked away and refreshed the page before
registering, the modal would pop up automatically. This commit changes
this behavior so that if a user is in a connected-and-unregistered
state, the prompt on the card changes to "Register to reveal". Clicking
this prompt displays the registration modal.

Finally, when a user is both connected and registered, the cards answer
can be revealed.