Sign registration transaction
This commit adds functionality to the registration modal that allows an
unregistered user to sign an EIP712-typed payload to permit DAI transfer
and register for the course.

The functionality to sign a transaction is wrapped in a hook that takes
a provider and the user's address. The hook returns a function that
takes the users nonce for the DAI contract and an expiry time for the
permit transaction. Upon completion of signing, the function returns an
object with properties v, r and s, to pass into the transaction.