Refactor flashcard/card.js for better readability
This commit contains a refactor of the card component to make the code a
bit clearer, more readable.

Generally, I changed things based on the following guidelines:
1. Move inline styles to a `styles` object at the bottom of the file if
   there are more than one or two attributes being added. Apply the
   style to the desired component via `styles.whateverYouNamedIt`
1. Define variables, functions, hooks, etc., grouped together by their
   function (e.g., hooks together, where possible) and by what they're
   related to, (e.g., style related variables)
1. Grouping `return`s together towards the end of the file where
   possible

There might be more we can do along the lines of extracting components
like @angelagilhotra did with the web3 modal, but these seem like a
great start.

Looking forward to hearing your thoughts

Co-authored-by: Anthony M. <anthony.moffa@hey.com>