Update registration flow to accommodate  scholarship registration (#336)
This commit adds additional functionality to the registration flow to
support scholarship registration. As an alternative to deposit one's own
DAI, a user will have the opportunity to register for the course via a
scholarship if one is available.

A new DeSchool contract was introduced on the backend that replaces and
extends the previous functionality of the KernelFactory contract. In
addition to the previously used `verify` and `permitAndRegister`
functions, the new contract also exposes `scholarshipAvailable` and
`registerScholar` functions.

When an unregistered user connects their wallet, a new check for an
available scholarship will run. If a scholarship is found, the modal
will register them for the course via the scholarship. If no scholarship
is available, the modal functions as it did previously, calling the
`permitAndRegister` function to deposit the user's DAI.

Co-authored-by: Anthony M <anthony.moffa@hey.com>