Fetch contract addresses based on current chainId
This commit enables automatic configuration of contract addresses based
on the current chainId. The constants file was split up into an object
for fetching contract ABIs and a function for fetching addresses.

The function for fetching addresses is set up so that if an unsupported,
is selected, we still return a valid value to avoid exceptions. It
returns mainnet addresses by default.