Create basic Redeem page
This commit introduces the structure for the redeem page where users
will be able to retrieve their deposit from the course.

Included are a number of new contract functions used to check the status
of a user's stake, the block in which they registered, the course length
and finally send a transaction to retrieve their deposit.

Still needed is the ability for a user to select whether or not they'd
like to retrieve their deposit or mint $LEARN instead.