Refactor contract calls to use wagmi hooks and integrate notificatins
This commit refactors our contract calls to use the wagmi hooks syntax.

Each of our contract calls were also rewritten as hooks to allow use to
stack the new notification functionality on top of the contract calls.

This lets us easily notifications plug into the `onError` and other
callback functions given on the wagmi `useContractRead` and
`useContractWrite` functions.

Another nice benefit of using the contract hooks is that we can avoid
having to state but being able to directly use the result of the call
instead of having to wait for an async result.