Support multiple AudioPlayer components per page
This commit enables multiple AudioPlayer instances for a given page.
Internally, the audio player code was referring to DOM elements by
static IDs, and so any additional audio players added to the page would
end up pointing at the first components DOM elements.

The fix here adds a UUID generation library to generate a UUID and
append it to the various DOM IDs that we use for handling audio-related
events.