Fix audio playback issue on iOS devices
There appears to be a restriction on iOS browsers that prevent audio
from being loaded and then played in a single async action. It looks
like the way React was setting up the audio in such a way that was
preventing the "play" button from working due to those restrictions.

To get around this, browsers with those restrictions need to see that
the audio-loading and play actions happens specifically due to captured
user clicks.

This commit sets up some special logic to check the user agent of the
current browser to see if the audio needs to be loaded specifically by
user action. To make this work more smoothly, we've dropped the `Audio`
element from the DOM in favor of constructing an Audio object in our
component code.

The automatic loading behavior for browsers that allow autoplay should
continue to work as before.